🐛 수정사항 반영 및 디테일 추가
