+++ src/interfaces/navInterface.ts
@@ -1,4 +1,3 @@
 export interface NavProps {
   userStatus: 'loggedOut' | 'generalUser' | 'adminUser';
-  setIsModalOpen: (isOpen: boolean) => void;
 }
